requirements.txt
 newspaper3k == 0.2.8
 fastapi == 0.92.0
 uvicorn == 0.20.0
 python-decouple == 3.8
 pydantic == 1.10.5
 apache-airflow == 2.5.1
 python-decouple == 3.8
 pymongo == 4.3.3
 python-dotenv==1.0.0
 requests==2.27.1
 beautifulsoup4==4.11.2
 requests-futures == 1.0.0